tools/kafka_consumer_adls/requirements.txt
confluent-kafka
azure-storage-file-datalake
pyyaml